+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Fixing DatafordelerBase missing configuration keys.
+
 ## [3.0.2] - 2025-10-28
 
 * [PR-27](https://github.com/OS2web/os2web_datalookup/pull/27)